// packages/vite/types.ts
interface BuildContext {
  outDir: string;
  packageName: string;
  isDev: boolean;
}
// ...
interface ServerBuildOptions {
  srcDir?: string;
  entry?: string;
  bundled?: string[];
  types?: boolean;
  nodeTarget?: string;
  plugins: Plugin[];
}